API/script/Chess_API/requirements.txt
pyyaml==6.0.*